cleaned up components
