+++ package.json
@@ -28,6 +28,7 @@
     "@radix-ui/react-scroll-area": "^1.2.4",
     "@radix-ui/react-separator": "^1.1.3",
     "@radix-ui/react-slot": "^1.2.0",
+    "@radix-ui/react-tooltip": "^1.2.7",
     "@t3-oss/env-core": "0.12.0",
     "@tanstack/react-query": "^5.72.2",
     "@webext-core/messaging": "^2.2.0",
